+++ forui/lib/src/widgets/breadcrumb.dart
@@ -49,7 +49,7 @@ class FBreadcrumb extends StatelessWidget {
 
   /// The divider placed between the children.
   ///
-  /// Defaults to [FLucideIcons.chevronRight].
+  /// Defaults to [FIcons.chevronRight].
   final Widget? divider;
 
   /// Creates an [FBreadcrumb].
@@ -58,7 +58,10 @@ class FBreadcrumb extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final style = this.style(context.theme.breadcrumbStyle);
-    final divider = IconTheme(data: style.iconStyle, child: this.divider ?? const Icon(FLucideIcons.chevronRight));
+    final divider = IconTheme(
+      data: style.iconStyle,
+      child: this.divider ?? context.theme.icons.chevronRight(context),
+    );
 
     return Row(
       children: [
@@ -125,6 +128,7 @@ abstract interface class FBreadcrumbItem extends Widget {
   /// displays a popover menu with the collapsed items.
   const factory FBreadcrumbItem.collapsed({
     required List<FItemGroup> menu,
+    Widget? icon,
     FPopoverMenuStyleDelta popoverMenuStyle,
     FPopoverControl popoverControl,
     ScrollController? scrollController,
@@ -160,6 +164,7 @@ abstract interface class FBreadcrumbItem extends Widget {
   /// displays a popover menu with the collapsed items.
   const factory FBreadcrumbItem.collapsedTiles({
     required List<FTileGroup> menu,
+    Widget? icon,
     FPopoverMenuStyleDelta popoverMenuStyle,
     FPopoverControl popoverControl,
     ScrollController? scrollController,
@@ -243,6 +248,7 @@ class _Crumb extends StatelessWidget implements FBreadcrumbItem {
 class _CollapsedCrumb extends StatefulWidget implements FBreadcrumbItem {
   final List<FTileGroup>? tileMenu;
   final List<FItemGroup>? itemMenu;
+  final Widget? icon;
   final FPopoverMenuStyleDelta popoverMenuStyle;
   final FPopoverControl popoverControl;
   final ScrollController? scrollController;
@@ -270,6 +276,7 @@ class _CollapsedCrumb extends StatefulWidget implements FBreadcrumbItem {
 
   const _CollapsedCrumb({
     required List<FItemGroup> menu,
+    this.icon,
     this.popoverMenuStyle = const .context(),
     this.popoverControl = const .managed(),
     this.scrollController,
@@ -300,6 +307,7 @@ class _CollapsedCrumb extends StatefulWidget implements FBreadcrumbItem {
 
   const _CollapsedCrumb.tiles({
     required List<FTileGroup> menu,
+    this.icon,
     this.popoverMenuStyle = const .context(),
     this.popoverControl = const .managed(),
     this.scrollController,
@@ -422,7 +430,10 @@ class _CollapsedCrumbState extends State<_CollapsedCrumb> with SingleTickerProvi
           onPress: _controller.toggle,
           child: Padding(
             padding: style.collapsedPadding,
-            child: IconTheme(data: style.iconStyle, child: const Icon(FLucideIcons.ellipsis)),
+            child: IconTheme(
+              data: style.iconStyle,
+              child: widget.icon ?? context.theme.icons.ellipsis(context),
+            ),
           ),
         ),
       );
@@ -455,7 +466,10 @@ class _CollapsedCrumbState extends State<_CollapsedCrumb> with SingleTickerProvi
           onPress: _controller.toggle,
           child: Padding(
             padding: style.collapsedPadding,
-            child: IconTheme(data: style.iconStyle, child: const Icon(FLucideIcons.ellipsis)),
+            child: IconTheme(
+              data: style.iconStyle,
+              child: widget.icon ?? context.theme.icons.ellipsis(context),
+            ),
           ),
         ),
       );