+++ forui/lib/src/widgets/select/content/content.dart
@@ -283,10 +283,11 @@ class FSelectContentStyle extends FPopoverStyle with Diagnosticable, _$FSelectCo
   FSelectContentStyle.inherit({
     required super.colors,
     required super.style,
+    required FIcons icons,
     required FTypography typography,
     required bool touch,
   }) : sectionStyle = .inherit(colors: colors, style: style, typography: typography, touch: touch),
-       scrollHandleStyle = .inherit(colors: colors, typography: typography),
+       scrollHandleStyle = .inherit(colors: colors, icons: icons, typography: typography),
        padding = const .symmetric(vertical: 6),
        super.inherit();
 }